Avoid UB when indexing into size-0 tensors
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65878

If we attempt to compute an offset into an empty tensor we trigger UB, since
we'd be adding an offset to a nullptr, which is UB
(https://reviews.llvm.org/D67122) even if we never use the pointer.

Since indexing into an empty tensor yields an empty tensor anyways, let's just
return the underlying (null) data ptr in this case.

ghstack-source-id: 139428615

Differential Revision: [D31296914](https://our.internmc.facebook.com/intern/diff/D31296914/)